feat: create a new user account using email and password Firebase Authentication
