+++ app/src/main/java/com/android/sample/MainActivity.kt
@@ -3,9 +3,6 @@ package com.android.sample
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
@@ -13,28 +10,29 @@ import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.semantics.testTag
 import androidx.compose.ui.tooling.preview.Preview
 import com.android.sample.resources.C
-import com.android.sample.ui.authentication.SignUpScreen
+import com.android.sample.ui.authentication.LoginScreen
 import com.android.sample.ui.theme.SampleAppTheme
 import com.google.firebase.FirebaseApp
+import com.google.firebase.auth.FirebaseAuth
 
 class MainActivity : ComponentActivity() {
 
-  // private lateinit var auth: FirebaseAuth
+  private lateinit var auth: FirebaseAuth
 
   override fun onCreate(savedInstanceState: Bundle?) {
     super.onCreate(savedInstanceState)
-      FirebaseApp.initializeApp(this)
+    FirebaseApp.initializeApp(this)
 
-    /*// Initialize Firebase Auth
+    // Initialize Firebase Auth
     auth = FirebaseAuth.getInstance()
     auth.currentUser?.let {
-        // Sign out the user if they are already signed in
-        // This is useful for testing purposes
-        auth.signOut()
-    }*/
+      // Sign out the user if they are already signed in
+      // This is useful for testing purposes
+      auth.signOut()
+    }
 
     setContent {
-      SignUpScreen()
+      LoginScreen()
       /*SampleAppTheme {
         // A surface container using the 'background' color from the theme
         Surface(
